Revert "Integrate Spark 2.0.0"
This reverts commit 279d95f0c2c32258d9a255cc20bbff8baaf139ca.
