+++ modules/common/src/main/pre-df/magic.scala
