+++ modules/common/src/main/scala/notebook/front/widgets/package.scala
@@ -18,7 +18,6 @@ import notebook.front.widgets.magic.SamplerImplicits._
 package object widgets
   extends Generic
   with Texts
-  with Images
   with Lists
   with Layouts
   with charts.Charts {
