+++ modules/spark/src/main/scala_2.11/spark-last/Boot.scala
@@ -6,12 +6,12 @@ import scala.tools.nsc.Settings
 import java.io.File
 
 object Boot {
-  /*def classServer(outputDir:File) = {
+  def classServer(outputDir:File) = {
     val conf = new SparkConf()
     val server = new HttpServer(conf, outputDir, new SecurityManager(conf))
     server
-  }*/
-
+  }
+  
   def createTempDir(
       root: String = System.getProperty("java.io.tmpdir"),
       namePrefix: String = "spark"): File = Utils.createTempDir(root, namePrefix)
