+++ modules/spark/src/main/scala_2.11/spark-last/HackSparkILoop.scala
@@ -13,15 +13,15 @@ import java.io.File
 
 import scala.tools.nsc.interpreter._
 
-class HackSparkILoop(out:JPrintWriter, val outputDir:File) extends org.apache.spark.repl.SparkILoop(None, out) {
+class HackSparkILoop(out:JPrintWriter, outputDir:File) extends org.apache.spark.repl.SparkILoop(None, out) {
 
   // note:
   // the creation of SecurityManager has to be lazy so SPARK_YARN_MODE is set if needed
-  /*val classServer = {
+  val classServer = {
     val s = org.apache.spark.Boot.classServer(outputDir)
     s.start
     s
-  }*/
+  }
 
   override def initializeSpark(): Unit = {
     // done using the metadata and init.sc
