--- modules/spark/src/main/scala_2.11/spark-pre2.0/HackIMain.scala
@@ -1,11 +1,0 @@
-package scala
-package tools.nsc
-package interpreter
-
-/**
- * Subclass to access some hidden things I need and also some custom behavior.
- */
-class HackIMain(settings: Settings, out: JPrintWriter) extends IMain(settings, out) {
-  def previousRequests = prevRequestList
-  override protected def parentClassLoader = settings.getClass.getClassLoader()
-}
