--- modules/spark/src/main/scala_2.11/spark-pre2.0/notebook/kernel/StringCompletorResolver.scala
@@ -1,12 +1,0 @@
-package notebook.kernel
-
-//import org.clapper.util.classutil.ClassUtil
-import notebook.util.StringCompletor
-
-object StringCompletorResolver {
-  lazy val completor = {
-    val className = "notebook.kernel.TestStringCompletor"
-    //ClassUtil.instantiateClass(className).asInstanceOf[StringCompletor]
-      Class.forName(className).getConstructor().newInstance().asInstanceOf[StringCompletor]
-  }
-}
